rename actionCreators to action-creators to fit naming standards for project
